+++ tests/force_control_test/force_control_test.ino
@@ -23,7 +23,7 @@ const float d = 0;
 // TrackRing objects
 TrackRing encoders[2];
 
-uint16_t clockSpeed = 1000000;
+uint32_t clockSpeed = 1000000;
 
 unsigned long timeStart = 0;
